atap 0.1.8
This is now a self-contained module with duplicated code from emulation
and utils to avoid that potential external users are depenendent on
github packages. A better solution may be to keep a self-contained copy
on matsim-libs and to keep using here the github packages dependencies.
Waiting with this until the pull request on matsim-libs has been
processed.